Separated Awakeness and Throwing responsibilities between two classes
