+++ Assets/Prefabs/Systems/---Systems---.prefab
@@ -30,7 +30,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 3534874787961472397}
-  m_RootOrder: 9
+  m_RootOrder: 8
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!114 &472111540
 MonoBehaviour:
@@ -94,7 +94,6 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children:
   - {fileID: 3534874787671726242}
-  - {fileID: 1664631557271132290}
   - {fileID: 7946989163454102114}
   - {fileID: 8071447776513924117}
   - {fileID: 3534874788697631547}
@@ -187,7 +186,7 @@ PrefabInstance:
     - target: {fileID: 5388189176525553835, guid: d578ddc6d06763e43b57dd20b1aa0b86,
         type: 3}
       propertyPath: m_RootOrder
-      value: 8
+      value: 7
       objectReference: {fileID: 0}
     - target: {fileID: 5388189176525553835, guid: d578ddc6d06763e43b57dd20b1aa0b86,
         type: 3}
@@ -312,91 +311,6 @@ Transform:
     type: 3}
   m_PrefabInstance: {fileID: 1266504443877479211}
   m_PrefabAsset: {fileID: 0}
---- !u!1001 &3534874787688001332
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    m_TransformParent: {fileID: 3534874787961472397}
-    m_Modifications:
-    - target: {fileID: 2744821312385824693, guid: ff4fef0c2fab35f478f33b46db963cce,
-        type: 3}
-      propertyPath: globalVolume
-      value: 
-      objectReference: {fileID: 0}
-    - target: {fileID: 2744821312385824693, guid: ff4fef0c2fab35f478f33b46db963cce,
-        type: 3}
-      propertyPath: playerController
-      value: 
-      objectReference: {fileID: 0}
-    - target: {fileID: 2744821312385824694, guid: ff4fef0c2fab35f478f33b46db963cce,
-        type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 2744821312385824694, guid: ff4fef0c2fab35f478f33b46db963cce,
-        type: 3}
-      propertyPath: m_LocalPosition.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 2744821312385824694, guid: ff4fef0c2fab35f478f33b46db963cce,
-        type: 3}
-      propertyPath: m_LocalPosition.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 2744821312385824694, guid: ff4fef0c2fab35f478f33b46db963cce,
-        type: 3}
-      propertyPath: m_LocalRotation.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 2744821312385824694, guid: ff4fef0c2fab35f478f33b46db963cce,
-        type: 3}
-      propertyPath: m_LocalRotation.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 2744821312385824694, guid: ff4fef0c2fab35f478f33b46db963cce,
-        type: 3}
-      propertyPath: m_LocalRotation.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 2744821312385824694, guid: ff4fef0c2fab35f478f33b46db963cce,
-        type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 2744821312385824694, guid: ff4fef0c2fab35f478f33b46db963cce,
-        type: 3}
-      propertyPath: m_RootOrder
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 2744821312385824694, guid: ff4fef0c2fab35f478f33b46db963cce,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 2744821312385824694, guid: ff4fef0c2fab35f478f33b46db963cce,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 2744821312385824694, guid: ff4fef0c2fab35f478f33b46db963cce,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 2744821312385824695, guid: ff4fef0c2fab35f478f33b46db963cce,
-        type: 3}
-      propertyPath: m_Name
-      value: AwakenessManager
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: ff4fef0c2fab35f478f33b46db963cce, type: 3}
---- !u!4 &1664631557271132290 stripped
-Transform:
-  m_CorrespondingSourceObject: {fileID: 2744821312385824694, guid: ff4fef0c2fab35f478f33b46db963cce,
-    type: 3}
-  m_PrefabInstance: {fileID: 3534874787688001332}
-  m_PrefabAsset: {fileID: 0}
 --- !u!1001 &3534874787961376712
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -458,7 +372,7 @@ PrefabInstance:
     - target: {fileID: 3076913251219283089, guid: c873d17681ed641ceb5b97b9fa56c366,
         type: 3}
       propertyPath: m_RootOrder
-      value: 6
+      value: 5
       objectReference: {fileID: 0}
     - target: {fileID: 3076913251219283089, guid: c873d17681ed641ceb5b97b9fa56c366,
         type: 3}
@@ -543,7 +457,7 @@ PrefabInstance:
     - target: {fileID: 4687665678606070099, guid: 65373c0d54cde75469cac3b6b80d27c0,
         type: 3}
       propertyPath: m_RootOrder
-      value: 3
+      value: 2
       objectReference: {fileID: 0}
     - target: {fileID: 4687665678606070099, guid: 65373c0d54cde75469cac3b6b80d27c0,
         type: 3}
@@ -633,7 +547,7 @@ PrefabInstance:
     - target: {fileID: 6865462128423233841, guid: 861584e79b77b2c4e883d6f2c7c47610,
         type: 3}
       propertyPath: m_RootOrder
-      value: 2
+      value: 1
       objectReference: {fileID: 0}
     - target: {fileID: 6865462128423233841, guid: 861584e79b77b2c4e883d6f2c7c47610,
         type: 3}
@@ -719,7 +633,7 @@ PrefabInstance:
     - target: {fileID: 8384848389310353534, guid: c95bf75eff27d7248a4c6b959b3d6d8c,
         type: 3}
       propertyPath: m_RootOrder
-      value: 5
+      value: 4
       objectReference: {fileID: 0}
     - target: {fileID: 8384848389310353534, guid: c95bf75eff27d7248a4c6b959b3d6d8c,
         type: 3}
@@ -822,7 +736,7 @@ PrefabInstance:
     - target: {fileID: 1521649277229448497, guid: d9f7fade4159d3b4d9cbbba1ab4cf76b,
         type: 3}
       propertyPath: m_RootOrder
-      value: 7
+      value: 6
       objectReference: {fileID: 0}
     - target: {fileID: 1521649277229448497, guid: d9f7fade4159d3b4d9cbbba1ab4cf76b,
         type: 3}
@@ -917,7 +831,7 @@ PrefabInstance:
     - target: {fileID: 7904942737515853286, guid: acf3f949c617bf741b1c993ee5c0a2d6,
         type: 3}
       propertyPath: m_RootOrder
-      value: 4
+      value: 3
       objectReference: {fileID: 0}
     - target: {fileID: 7904942737515853286, guid: acf3f949c617bf741b1c993ee5c0a2d6,
         type: 3}
