+++ Assets/Scripts/AwakenessManager.cs
@@ -1,9 +1,16 @@
-﻿using UnityEngine;
+﻿using System.Collections.Generic;
+using UnityEngine;
 using UnityEngine.Rendering;
 using UnityEngine.Rendering.Universal;
 
 public class AwakenessManager : Singleton<AwakenessManager>
 {
+    [Header("Awakeness Calculation")]
+    public float awakenessIncrease = 0.01f;
+    public float dangerRadius = 50.0f;
+    
+    private List<Coroutine> spawnedCoroutines;
+    
     [Header("Post Processing")]
     public Volume globalVolume;
     public float vignetteMultiplier = 0.45f;
@@ -29,6 +36,8 @@ public class AwakenessManager : Singleton<AwakenessManager>
 
     private void Start()
     {
+        spawnedCoroutines = new List<Coroutine>();
+        
         globalVolume.profile.TryGet(out vignette);
         globalVolume.profile.TryGet(out motionBlur);
         globalVolume.profile.TryGet(out filmGrain);
@@ -46,6 +55,58 @@ public class AwakenessManager : Singleton<AwakenessManager>
         movementController = GameManager.Instance.GetMovementController();
         UIManager.Instance.staminaBar.OnChange += UpdateMovementBuffs;
     }
+    
+    private void FixedUpdate()
+    {
+        float minDistance = DistToClosestEnemy();
+
+        if (minDistance < dangerRadius)
+        {
+            float awakenessGain = Mathf.Pow(((dangerRadius - minDistance) / dangerRadius), 2.0f);
+            HandleIncreaseAwakeness(awakenessGain);
+        }
+    }
+    
+    private float DistToClosestEnemy()
+    {
+        GameObject[] enemies = GameObject.FindGameObjectsWithTag("Enemy");
+        float minDistance = Mathf.Infinity;
+        Vector3 position = transform.position;
+        foreach (GameObject enemy in enemies)
+        {
+            float curDistance = Vector3.Distance(enemy.transform.position, position);
+            if (curDistance < minDistance)
+            {
+                minDistance = curDistance;
+            }
+        }
+        return minDistance;
+    }
+    
+    void HandleIncreaseAwakeness(float multiplier)
+    {
+        spawnedCoroutines.Add(StartCoroutine(UIManager.Instance.staminaBar.IncreaseFillBy(multiplier * awakenessIncrease)));
+    }
+    
+    // Note: Might need to do more testing if this is actually doing anything considerable...
+    //  but better safe than sorry to make sure coroutines we spawn are no longer running when we enter a minigame
+    void StopAllSpawnedCoroutines()
+    {
+        // No loose coroutines in MY house!
+        foreach (Coroutine c in spawnedCoroutines)
+        {
+            StopCoroutine(c);
+        }
+        spawnedCoroutines.Clear();
+    }
+
+    private void OnDisable()
+    {
+        StopAllSpawnedCoroutines();
+    }
+    
+    
+    // // //  AWAKENESS LISTENERS // // //
 
     private float GetInterpolatedFillAmount(float fillAmount)
     {
